[Socket] Move Worker out of Socket_2_1.cs to SocketAsyncWorker.cs
